// src/types/recipe.ts
export interface Recipe {
    calories: number;
    category: string;
    cook_time: number;
    cost: number;
    created_at: string;
    created_by: string;
    description: string;
    disclaimer: string;
    id: string;
    image_url: string;
    instructions: string;
    name: string;
    prep_time: number;
    published: boolean;
    servings: number;
    when_to_eat: string;
}